Put InMemory + Persistent repositories in separate files
