+++ lib/data_mapper/repository.rb
@@ -85,66 +85,5 @@ module DataMapper
       @map[name] = build(name, Veritas::Relation::Header.coerce(header))
       self
     end
-
-    # A repository backed by in memory relations
-    #
-    # @api private
-    class InMemory < self
-      private
-
-      # Build veritas in memory relation
-      #
-      # @param [Symbol] name
-      #   the relation name
-      #
-      # @param [Veritas::Relation::Header] header
-      #
-      # @return [Veritas::Relation]
-      #
-      # @api private
-      #
-      def build(_name, header)
-        Veritas::Relation.new(header, [])
-      end
-    end # class InMemory
-
-    # A persistent repository backed by base relations accessible via an adapter
-    #
-    # @api private
-    class Persistent < self
-
-      # Initialize a new instance
-      #
-      # @param [#to_sym] name
-      #   the repository's name
-      #
-      # @param [Veritas::Adapter] adapter
-      #   the adapter to access relations
-      #
-      # @return [undefined]
-      #
-      # @api private
-      def initialize(name, adapter)
-        super(name)
-        @adapter = adapter
-      end
-
-      private
-
-      # Build veritas gateway relation
-      #
-      # @param [Symbol] name
-      #   the relatio name
-      #
-      # @param [Veritas::Relation::Header] header
-      #
-      # @return [Veritas::Adapter::Gateway]
-      #
-      # @api private
-      #
-      def build(name, header)
-        @adapter.gateway(Veritas::Relation::Base.new(name, header))
-      end
-    end # class Persistent
   end # class Repository
 end # module DataMapper
